Implement most basic of OpenGL rendering
All OpenGL code is now housed inside of the platform package. For
simplicity and following some tutorials the default "cube" is now a
screen-space triangle. I've set up a test vertex and fragment shader and
figured out the whole deal with VAO and bindings. Now the app, when run,
displays a red triangle on black background, yay!