Move to a RenderQueue / RenderOperation setup
The overall structure I want here is to build up a structure of objects
(here named RenderQueue, which has many RenderOperations) that define
everything to be rendered for the next frame. Then this object can be
thrown at the renderer implementation and nothing else needs to be done
on the app side, letting game code tick over once again.