Save the default material to the unfound material name
Get is called every frame when choosing a material to render for the
entity in question, so without this memoization it was spamming logs. We
can assume that if a material is not defined it probably won't be, but
the system is setup so that it can be defined later during runtime.