Delete entities and clean up appropriately
Entities can now be deleted from the system and the appropriate
Listeners will be told to clean up for this Entity if any such clean up
must take place.

Along with this set up an Update call to EntityDB which will do an
initial run through of any dirty Entities at the beginning of each frame
to make requested changes from the previous frame.