Flag entity component removal to be cleaned up next frame
Turns out this gets slightly complicated and I'll try to do some
refactoring here. Basically, removing a Component from an Entity needs
to tell all listeners that did have the Entity because of this component
to clean up the Entity. However the listeners still need access to that
removed Component when doing cleanup so we can't just nil things out.

I've instead gone the path (and updated Add Component to match) of
flagging the changes, updating important methods to know about removed
components, and adding a private callback to finally clear out the
component from the map.