Give Entities a pointer back to the DB that it's registered with
The async add/delete/manage cycle ended up being a big pain to get all
the cases working right (specifically, replacing an entity with one of
the same type would have turned into a right mess). After putting this
off and trying to work without a direct linkage between Entity and it's
DB, I gave in and am now going this route.

Entities will now inform their DB of any component changes, as well as
their own destruction. I will have to keep an eye on performance if
these calls get run a lot (particularly for graphics) to see if some of
this needs to be tasked out.