Poll for requested input events
There are cases where the game requires per-frame polling for events.
This update opens up an API to specify which events should be polled for
(if any key linked to an event is Pressed, the Event fires) and hooks
the polling check into RecentEvents, so that if nothing is asking for
input, no polling has to happen.